pyproject.toml
[project]
name = "energymatchingeqprop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "absl-py>=2.4.0",
    "einops>=0.8.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "omegaconf>=2.3.0",
    "pandas>=3.0.0",
    "pot>=0.9.6.post1",
    "python-levenshtein>=0.27.3",
    "torch>=2.10.0",
    "torch-fidelity>=0.3.0",
    "torchcfm>=1.0.7",
    "torchdyn>=1.0.6",
    "torchmetrics>=1.8.2",
    "torchsde>=0.2.6",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
]
